Pairwise/pyproject.toml
[project]
name = "pairwise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.120.4",
    "fastmcp>=2.13.0.2",
    "google-generativeai>=0.8.5",
    "joblib>=1.5.2",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "prophet>=1.2.1",
    "requests>=2.32.5",
]
